Finalización del proyecto y documentación

+++ Controllers/Categorias.php
@@ -66,11 +66,28 @@ class Categorias extends Controller{
         for($i=0; $i < count($data); $i++){
           
                 $data[$i]['estado'] = '<span class="badge badge-success">Activo</span>';
-                $data[$i]['acciones'] = '<div>            
-                <button type="button" class="btn btn-primary" onclick="editarCategoria('.$data[$i]['id'].');" title="Editar"><i class="fas fa-edit"></i></button>   
-                <button type="button" class="btn btn-danger" onclick="eliminarCategoria('.$data[$i]['id'].');" title="Eliminar"><i class="far fa-trash-alt"></i></button>            
-               </div>';
+                if($_SESSION['rol'] == 'Administrador'){
+                    $data[$i]['acciones'] = '<div>            
+                    <button type="button" class="btn btn-primary" onclick="editarCategoria('.$data[$i]['id'].');" title="Editar"><i class="fas fa-edit"></i></button>   
+                    <button type="button" class="btn btn-danger" onclick="eliminarCategoria('.$data[$i]['id'].');" title="Eliminar"><i class="far fa-trash-alt"></i></button>            
+                </div>';
                                 
+                }else if( $_SESSION['rol'] == 'Supervisor'){
+                    $data[$i]['acciones'] = '<div>            
+                    <button type="button" class="btn btn-primary" onclick="editarCategoria('.$data[$i]['id'].');" title="Editar"><i class="fas fa-edit"></i></button>   
+                    <button type="button"  disabled="" class="btn btn-danger" onclick="eliminarCategoria('.$data[$i]['id'].');" title="Eliminar"><i class="far fa-trash-alt"></i></button>            
+                </div>';
+                }else{
+   
+                         $data[$i]['acciones'] = '<div>            
+                    <button type="button" disabled="" class="btn btn-primary" onclick="editarCategoria('.$data[$i]['id'].');" title="Editar"><i class="fas fa-edit"></i></button>   
+                    <button type="button" disabled="" class="btn btn-danger" onclick="eliminarCategoria('.$data[$i]['id'].');" title="Eliminar"><i class="far fa-trash-alt"></i></button>            
+                   </div>';
+           
+                   
+                                    
+                }
+     
         }      
         echo json_encode($data, JSON_UNESCAPED_UNICODE);
         die();
@@ -109,6 +126,17 @@ class Categorias extends Controller{
             $msg = (array('eliminado'=>true, 'post' => 'El Categoria fue eliminado con éxito.'));
         }else{
             $msg = (array('eliminado'=>false, 'msg' => 'Error al eliminar el Categoria.'));
+        }
+        echo json_encode($msg, JSON_UNESCAPED_UNICODE);
+        die();
+    }
+    public function vaciarCategorias(){
+        $data = $this->model->vaciarCategorias();  
+
+        if($data == 1){
+            $msg = (array('eliminado'=>true, 'post' => 'Las Categorias fueron vaciados con éxito.'));
+        }else{
+            $msg = (array('eliminado'=>false, 'msg' => 'Error al vaciar las Categoria.'));
         }
         echo json_encode($msg, JSON_UNESCAPED_UNICODE);
         die();
